+++ cache.go
@@ -2,37 +2,102 @@ package sdpcache
 
 import (
 	"context"
+	"crypto/sha1"
+	"errors"
 	"fmt"
-	"math"
 	"strings"
 	"sync"
 	"time"
 
+	"github.com/google/btree"
 	"github.com/overmindtech/sdp-go"
-	log "github.com/sirupsen/logrus"
 )
 
-// MinWaitDefault The default minimum wait time
-const MinWaitDefault = (5 * time.Second)
-
-// Cache is responsible for caching data between calls
-type Cache struct {
-	// Name used for logging since there are many caches
-	Name    string
-	Storage []CachedResult
-
-	// The minimum amount of time to wait between purges, defaults to
-	// MinWaitDefault
-	MinWaitTime time.Duration
-
-	// Mutexes that are used to ensure thread safety of the underlying storage
-	storageMutex sync.RWMutex
-	context      context.Context
-	cancel       context.CancelFunc
+type IndexValues struct {
+	SSTHash              SSTHash
+	UniqueAttributeValue string
+	Method               sdp.RequestMethod
+	Query                string
 }
 
-// Tags A map of key-value pairs that can be searched on
-type Tags map[string]string
+type CacheQuery struct {
+	SST                  SST // *required
+	UniqueAttributeValue *string
+	Method               *sdp.RequestMethod
+	Query                *string
+}
+
+func (cc CacheQuery) String() string {
+	fields := []string{
+		("SourceName=" + cc.SST.SourceName),
+		("Scope=" + cc.SST.Scope),
+		("Type=" + cc.SST.Type),
+	}
+
+	if cc.UniqueAttributeValue != nil {
+		fields = append(fields, ("UniqueAttributeValue=" + *cc.UniqueAttributeValue))
+	}
+
+	if cc.Method != nil {
+		fields = append(fields, ("Method=" + cc.Method.String()))
+	}
+
+	if cc.Query != nil {
+		fields = append(fields, ("Query=" + *cc.Query))
+	}
+
+	return strings.Join(fields, ", ")
+}
+
+// Matches Returns whether or not the supplied index values match the
+// CacheQuery, excluding the SST since this will have already been validated.
+// Note that this only checks values that ave actually been set in the
+// CacheQuery
+func (cc CacheQuery) Matches(i IndexValues) bool {
+	// Check for any mismatches on the values that are set
+	if cc.Method != nil {
+		if *cc.Method != i.Method {
+			return false
+		}
+	}
+
+	if cc.Query != nil {
+		if *cc.Query != i.Query {
+			return false
+		}
+	}
+
+	if cc.UniqueAttributeValue != nil {
+		if *cc.UniqueAttributeValue != i.UniqueAttributeValue {
+			return false
+		}
+	}
+
+	return true
+}
+
+var ErrCacheNotFound = errors.New("not found in cache")
+
+// SST A combination of SourceName, Scope and Type, all of which must be
+// provided
+type SST struct {
+	SourceName string
+	Scope      string
+	Type       string
+}
+
+// Hash Creates a new SST hash from a given SST
+func (s SST) Hash() SSTHash {
+	h := sha1.New()
+	h.Write([]byte(s.SourceName))
+	h.Write([]byte(s.Scope))
+	h.Write([]byte(s.Type))
+
+	sum := make([]byte, 0)
+	sum = h.Sum(sum)
+
+	return SSTHash(fmt.Sprintf("%x", sum))
+}
 
 // CachedResult An item including cache metadata
 type CachedResult struct {
@@ -42,215 +107,357 @@ type CachedResult struct {
 	// Error is the error that we want
 	Error error
 
-	// CacheDuration How long to cache the item for
-	CacheDuration time.Duration
+	// The time at which this item expires
+	Expiry time.Time
 
-	// InsertTime When the item was inserted into the cache
-	InsertTime time.Time
-
-	// Tags are arbitrary strings maps that can be searched on
-	Tags Tags
+	// Values that we use for calculating indexes
+	IndexValues IndexValues
 }
 
-// Expired Tells us if the item's cache has expired
-func (cr *CachedResult) Expired() bool {
-	return cr.InsertTime.Add(cr.CacheDuration).Before(time.Now())
+// SSTHash Represents the hash of `SourceName`, `Scope` and `Type`
+type SSTHash string
+
+type Cache struct {
+	// Minimum amount of time to wait between cache purges
+	MinWaitTime time.Duration
+
+	// The timer that is used to trigger the next purge
+	purgeTimer *time.Timer
+
+	// The time that the purger will run next
+	nextPurge time.Time
+
+	indexes map[SSTHash]*indexSet
+
+	// This index is used to track item expiries, since items can have different
+	// expiry durations we need to use a btree here rather than just appending
+	// to a slice or something. The purge process uses this to determine what
+	// needs deleting, then calls into each specific index to delete as required
+	expiryIndex *btree.BTreeG[*CachedResult]
+
+	// Mutex for reading caches
+	indexMutex sync.RWMutex
+
+	// Ensures that purge stats like `purgeTimer` and `nextPurge` aren't being
+	// modified concurrently
+	purgeMutex sync.Mutex
 }
 
-// HasTag returns a boolean if the CachedResult has a given tag with a given
-// value
-func (cr *CachedResult) HasTag(key string, value string) bool {
-	// Range over the tags that the result has
-	for tk, tv := range cr.Tags {
-		// Return true as soon as we find a match
-		if tk == key && tv == value {
-			return true
-		}
-	}
-
-	// If no match was found return false
-	return false
-}
-
-// HasTags returns whether a given result has all of the tags supplied. This is
-// a logical AND query
-func (cr *CachedResult) HasTags(tags Tags) bool {
-	// Range over all the required sets of tags
-	for tk, tv := range tags {
-		// Check if the result has the required tag, immediately return false if
-		// not, otherwise continue on and check all of the other tags
-		if !cr.HasTag(tk, tv) {
-			return false
-		}
-	}
-
-	return true
-}
-
-// CacheNotFoundError is returned when an item is not found in the cache.
-// Usually this would be handled very gracefully
-type CacheNotFoundError Tags
-
-// StoreItem Store an item in the cache.
-//
-// Note that all stored items will have the following default tags applied:
-//
-//	type: (item.Type)
-//	scope: (item.Scope)
-//	uniqueAttributeValue: (item.Attributes[item.UniqueAttribute])
-//
-// These default tags mean that each full tag set will be unique to an item
-func (c *Cache) StoreItem(item *sdp.Item, duration time.Duration, tags Tags) {
-	var allTags Tags
-	var itemCopy sdp.Item
-
-	allTags = addDefaultItemTags(item, tags)
-
-	// Lock to ensure thread safety
-	c.storageMutex.Lock()
-	defer c.storageMutex.Unlock()
-
-	if c.unsafeExists(allTags) {
-		// If the item is already in the cache then delete it since we have a
-		// new version
-		c.unsafeDelete(allTags)
-	}
-
-	// Clone the item for storage
-	item.Copy(&itemCopy)
-
-	si := CachedResult{
-		Item:          &itemCopy,
-		Error:         nil,
-		CacheDuration: duration,
-		InsertTime:    time.Now(),
-		Tags:          addDefaultItemTags(item, tags),
-	}
-
-	c.Storage = append(c.Storage, si)
-}
-
-// BulkStoreItem stores many items at once instead of just one
-func (c *Cache) BulkStoreItem(items []*sdp.Item, duration time.Duration, tags Tags) {
-	for _, item := range items {
-		c.StoreItem(item, duration, tags)
+func NewCache() *Cache {
+	return &Cache{
+		indexes: make(map[SSTHash]*indexSet),
+		expiryIndex: btree.NewG(2, func(a, b *CachedResult) bool {
+			return a.Expiry.Before(b.Expiry)
+		}),
 	}
 }
 
-// StoreError Store an item in the cache.
-//
-// Note that all errors will get the following tags by default:
-//
-//	isError: true
-func (c *Cache) StoreError(err error, duration time.Duration, tags Tags) {
-	// Lock to ensure thread safety
-	c.storageMutex.Lock()
-	defer c.storageMutex.Unlock()
-
-	// Ensure that the error hasn't already been stored in the cache. Since
-	// error uniqueness isn't determined in any good way we will use the tags
-	// passed in here
-	c.unsafeDelete(addDefaultErrorTags(err, tags))
-
-	si := CachedResult{
-		Item:          &sdp.Item{},
-		Error:         err,
-		CacheDuration: duration,
-		InsertTime:    time.Now(),
-		Tags:          addDefaultErrorTags(err, tags),
-	}
-
-	c.Storage = append(c.Storage, si)
+type indexSet struct {
+	uniqueAttributeValueIndex *btree.BTreeG[*CachedResult]
+	methodIndex               *btree.BTreeG[*CachedResult]
+	queryIndex                *btree.BTreeG[*CachedResult]
 }
 
-// Search will find all items with a given set of tags. If multiple tags
-// are supplied an AND is used. If a cached error is found with matching tags it
-// will be returned immediately
-func (c *Cache) Search(searchTags Tags) ([]*sdp.Item, error) {
-	c.storageMutex.RLock()
-	defer c.storageMutex.RUnlock()
-
-	// Call unsafeSearch, but within the read lock that we just acquired
-	return c.unsafeSearch(searchTags)
+func newIndexSet() *indexSet {
+	return &indexSet{
+		uniqueAttributeValueIndex: btree.NewG(2, func(a, b *CachedResult) bool {
+			return sortString(a.IndexValues.UniqueAttributeValue, a.Item) < sortString(b.IndexValues.UniqueAttributeValue, b.Item)
+		}),
+		methodIndex: btree.NewG(2, func(a, b *CachedResult) bool {
+			return sortString(a.IndexValues.Method.String(), a.Item) < sortString(b.IndexValues.Method.String(), b.Item)
+		}),
+		queryIndex: btree.NewG(2, func(a, b *CachedResult) bool {
+			return sortString(a.IndexValues.Query, a.Item) < sortString(b.IndexValues.Query, b.Item)
+		}),
+	}
 }
 
-// unsafeSearch Is the backend search method, but does not handle locking of the
-// storage. Only use this method when you are handing the locking in a higher
-// layer to avoid race conditions
-func (c *Cache) unsafeSearch(searchTags Tags) ([]*sdp.Item, error) {
-	results := make([]*sdp.Item, 0)
+// Search Runs a given query against the cache. If a cached error is found it
+// will be returned immediately, if nothing is found a ErrCacheNotFound will
+// be returned. Otherwise this will return items that match ALL of the given
+// query parameters
+func (c *Cache) Search(cc CacheQuery) ([]*sdp.Item, error) {
+	items := make([]*sdp.Item, 0)
 
-	for _, r := range c.Storage {
-		if r.HasTags(searchTags) {
-			var returnItem sdp.Item
-
-			// If there is an error then we want that returned immediately
-			if r.Error != nil {
-				return results, r.Error
-			}
-
-			// Duplicate the item by value
-			r.Item.Copy(&returnItem)
-
-			// Return a pointer to the new duplicate. This should mean that
-			// things like attaching linked items to something that was returned
-			// from the cache shouldn't affect the cached items as they are no
-			// longer pointers to the same objects in memory
-			results = append(results, &returnItem)
-		}
-	}
+	results := c.getResults(cc)
 
 	if len(results) == 0 {
-		return make([]*sdp.Item, 0), CacheNotFoundError(searchTags)
+		return nil, ErrCacheNotFound
 	}
 
-	return results, nil
+	// If there is an error we want to return that, so we need to range over the
+	// results and separate items and errors. This is computationally less
+	// efficient than extracting errors inside of `getResults()` but logically
+	// it's a lot less complicated since `Delete()` uses the same method but
+	// applies different logic
+	for _, res := range results {
+		if res.Error != nil {
+			return nil, res.Error
+		}
+
+		// Return a copy of the item so the user can do whatever they want with
+		// it
+		itemCopy := sdp.Item{}
+		res.Item.Copy(&itemCopy)
+
+		items = append(items, &itemCopy)
+	}
+
+	return items, nil
 }
 
-// unsafeExists Returns whether or not anything exists with the given tags, does
-// not handle storage locking
-func (c *Cache) unsafeExists(searchTags Tags) bool {
-	for _, r := range c.Storage {
-		if r.HasTags(searchTags) {
+// Delete Deletes anything that matches the given cache query
+func (c *Cache) Delete(cc CacheQuery) {
+	c.deleteResults(c.getResults(cc))
+}
+
+// getResults Searches indexes for cached results, doing no other logic. If
+// nothing is found an empty slice will be returned.
+func (c *Cache) getResults(cc CacheQuery) []*CachedResult {
+	c.indexMutex.RLock()
+	defer c.indexMutex.RUnlock()
+
+	results := make([]*CachedResult, 0)
+
+	// Get the relevant set of indexes based on the SST Hash
+	sstHash := cc.SST.Hash()
+	indexes, exists := c.indexes[sstHash]
+	pivot := CachedResult{
+		IndexValues: IndexValues{
+			SSTHash: sstHash,
+		},
+	}
+
+	if !exists {
+		// If we don't have a set of indexes then it definitely doesn't exist
+		return results
+	}
+
+	// Start with the most specific index and fall back to the least specific.
+	// Checking all matching items and returning. These is no need to check all
+	// indexes since they all have the same content
+	if cc.UniqueAttributeValue != nil {
+		pivot.IndexValues.UniqueAttributeValue = *cc.UniqueAttributeValue
+
+		indexes.uniqueAttributeValueIndex.AscendGreaterOrEqual(&pivot, func(result *CachedResult) bool {
+			if *cc.UniqueAttributeValue == result.IndexValues.UniqueAttributeValue {
+				if cc.Matches(result.IndexValues) {
+					results = append(results, result)
+				}
+
+				// Always return true so that we continue to iterate
+				return true
+			}
+
+			return false
+		})
+
+		return results
+	}
+
+	if cc.Query != nil {
+		pivot.IndexValues.Query = *cc.Query
+
+		indexes.queryIndex.AscendGreaterOrEqual(&pivot, func(result *CachedResult) bool {
+			if *cc.Query == result.IndexValues.Query {
+				if cc.Matches(result.IndexValues) {
+					results = append(results, result)
+				}
+
+				// Always return true so that we continue to iterate
+				return true
+			}
+
+			return false
+		})
+
+		return results
+	}
+
+	if cc.Method != nil {
+		pivot.IndexValues.Method = *cc.Method
+
+		indexes.methodIndex.AscendGreaterOrEqual(&pivot, func(result *CachedResult) bool {
+			if *cc.Method == result.IndexValues.Method {
+				// If the methods match, check the rest
+				if cc.Matches(result.IndexValues) {
+					results = append(results, result)
+				}
+
+				// Always return true so that we continue to iterate
+				return true
+			}
+
+			return false
+		})
+
+		return results
+	}
+
+	// If nothing other than SST has been set then return everything
+	indexes.methodIndex.Ascend(func(result *CachedResult) bool {
+		results = append(results, result)
+
+		return true
+	})
+
+	return results
+}
+
+// StoreItem Stores an item in the cache. Note that this item must be fully
+// populated (including metadata) for indexing to work correctly
+func (c *Cache) StoreItem(item *sdp.Item, duration time.Duration) {
+	if item == nil || c == nil {
+		return
+	}
+
+	itemCopy := sdp.Item{}
+	item.Copy(&itemCopy)
+
+	res := CachedResult{
+		Item:   &itemCopy,
+		Error:  nil,
+		Expiry: time.Now().Add(duration),
+		IndexValues: IndexValues{
+			UniqueAttributeValue: itemCopy.UniqueAttributeValue(),
+			Method:               itemCopy.Metadata.SourceRequest.Method,
+			Query:                itemCopy.Metadata.SourceRequest.Query,
+			SSTHash: SST{
+				SourceName: itemCopy.Metadata.SourceName,
+				Scope:      itemCopy.Scope,
+				Type:       itemCopy.Type,
+			}.Hash(),
+		},
+	}
+
+	c.storeResult(res)
+}
+
+// StoreError Stores an error for the given duration. Since we can't determine
+// the index values from the error itself, the user also needs to pass these in.
+func (c *Cache) StoreError(err error, duration time.Duration, indexValues IndexValues) {
+	if c == nil || err == nil {
+		return
+	}
+
+	res := CachedResult{
+		Item:        nil,
+		Error:       err,
+		Expiry:      time.Now().Add(duration),
+		IndexValues: indexValues,
+	}
+
+	c.storeResult(res)
+}
+
+func (c *Cache) storeResult(res CachedResult) {
+	c.indexMutex.Lock()
+	defer c.indexMutex.Unlock()
+
+	// Create the index if it doesn't exist
+	indexes, ok := c.indexes[res.IndexValues.SSTHash]
+
+	if !ok {
+		indexes = newIndexSet()
+		c.indexes[res.IndexValues.SSTHash] = indexes
+	}
+
+	// Add the item to the indexes
+	indexes.methodIndex.ReplaceOrInsert(&res)
+	indexes.queryIndex.ReplaceOrInsert(&res)
+	indexes.uniqueAttributeValueIndex.ReplaceOrInsert(&res)
+
+	// Add the item ot the expiry index
+	c.expiryIndex.ReplaceOrInsert(&res)
+
+	// Update the purge time if required
+	c.setNextPurgeIfNewer(res.Expiry)
+}
+
+// sortString Returns the string that the cached result should be sorted on.
+// This has a prefix of the index value and suffix of the GloballyUniqueName if
+// relevant
+func sortString(indexValue string, item *sdp.Item) string {
+	if item == nil {
+		return indexValue
+	} else {
+		return indexValue + item.GloballyUniqueName()
+	}
+}
+
+// PurgeStats Stats about the Purge
+type PurgeStats struct {
+	// How many items were timed out of the cache
+	NumPurged int
+	// How long the purging took overall
+	TimeTaken time.Duration
+	// The expiry time of the next item to expire. If there are no more items in
+	// the cache, this will be nil
+	NextExpiry *time.Time
+}
+
+// deleteResults Deletes many cached results at once
+func (c *Cache) deleteResults(results []*CachedResult) {
+	c.indexMutex.Lock()
+	defer c.indexMutex.Unlock()
+
+	for _, res := range results {
+		if indexSet, ok := c.indexes[res.IndexValues.SSTHash]; ok {
+			// For each expired item, delete it from all of the indexes that it will be in
+			if indexSet.methodIndex != nil {
+				indexSet.methodIndex.Delete(res)
+			}
+			if indexSet.queryIndex != nil {
+				indexSet.queryIndex.Delete(res)
+			}
+			if indexSet.uniqueAttributeValueIndex != nil {
+				indexSet.uniqueAttributeValueIndex.Delete(res)
+			}
+		}
+
+		c.expiryIndex.Delete(res)
+	}
+}
+
+// Purge Purges all expired items from the cache. The user must pass in the
+// `before` time. All items that expired before this will be purged. Usually
+// this would be just `time.Now()` however it could be overridden for testing
+func (c *Cache) Purge(before time.Time) PurgeStats {
+	// Store the current time rather than calling it a million times
+	start := time.Now()
+
+	var nextExpiry *time.Time
+
+	expired := make([]*CachedResult, 0)
+
+	// Look through the expiry cache and work out what has expired
+	c.indexMutex.RLock()
+	c.expiryIndex.Ascend(func(res *CachedResult) bool {
+		if res.Expiry.Before(before) {
+			expired = append(expired, res)
+
 			return true
 		}
+
+		// Take note of the next expiry so we can schedule the next run
+		nextExpiry = &res.Expiry
+
+		// As soon as hit this we'll stop ascending
+		return false
+	})
+	c.indexMutex.RUnlock()
+
+	c.deleteResults(expired)
+
+	return PurgeStats{
+		NumPurged:  len(expired),
+		TimeTaken:  time.Since(start),
+		NextExpiry: nextExpiry,
 	}
-
-	return false
 }
 
-// Delete Removes items form the cache that have matching tags
-func (c *Cache) Delete(searchTags Tags) error {
-	c.storageMutex.Lock()
-	defer c.storageMutex.Unlock()
-
-	// Use a write lock, then call the unsafe method
-	return c.unsafeDelete(searchTags)
-}
-
-// unsafeDelete Deletes a set if items from the cache using tags, but without
-// locking. When using this function you need to handle the locking yourself
-func (c *Cache) unsafeDelete(searchTags Tags) error {
-	newStorage := make([]CachedResult, 0)
-
-	for _, r := range c.Storage {
-		if !r.HasTags(searchTags) {
-			newStorage = append(newStorage, r)
-		}
-	}
-
-	// Write the new data
-	c.Storage = newStorage
-
-	return CacheNotFoundError(searchTags)
-}
-
-// Clear Clears all items from the cache
-func (c *Cache) Clear() {
-	c.storageMutex.Lock()
-	c.Storage = make([]CachedResult, 0)
-	c.storageMutex.Unlock()
-}
+// MinWaitDefault The default minimum wait time
+const MinWaitDefault = (5 * time.Second)
 
 // GetMinWaitTime Returns the minimum wait time or the default if not set
 func (c *Cache) GetMinWaitTime() time.Duration {
@@ -261,182 +468,77 @@ func (c *Cache) GetMinWaitTime() time.Duration {
 	return c.MinWaitTime
 }
 
-// DefaultItemTags Returns default tags for a given item
-func DefaultItemTags(item *sdp.Item) Tags {
-	return Tags{
-		"type":                 item.Type,
-		"scope":                item.Scope,
-		"uniqueAttributeValue": item.UniqueAttributeValue(),
-	}
-}
-
-// DefaultErrorTags Returns default tags for any error
-func DefaultErrorTags(e error) Tags {
-	return Tags{
-		"isError":    "true",
-		"errorValue": e.Error(),
-	}
-}
-
-// addDefaultItemTags Appends default tags
-func addDefaultItemTags(item *sdp.Item, tags Tags) Tags {
-	return mergeTags(tags, DefaultItemTags(item))
-}
-
-// addDefaultErrorTags Appends default tags
-func addDefaultErrorTags(e error, tags Tags) Tags {
-	return mergeTags(tags, DefaultErrorTags(e))
-}
-
-func mergeTags(aTags Tags, bTags Tags) Tags {
-	endTags := make(Tags)
-
-	for k, v := range aTags {
-		endTags[k] = v
-	}
-
-	for k, v := range bTags {
-		endTags[k] = v
-	}
-
-	return endTags
-}
-
-func (c CacheNotFoundError) Error() string {
-	var tagPairs []string
-	var tagString string
-
-	for k, v := range c {
-		tagPairs = append(tagPairs, (k + " = " + v))
-	}
-
-	tagString = strings.Join(tagPairs, ", ")
-
-	return fmt.Sprintf("cache: Cannot find value with tags %v", tagString)
-}
-
-// PurgeStats Stats about the Purge
-type PurgeStats struct {
-	// How many items were timed out of the cache
-	NumPurged int
-	// How long the Purgeing took overall
-	TimeTaken time.Duration
-	// Shortest cache duration of the remaining items
-	ShortestCacheRemaining time.Duration
-}
-
-// Purge Prunes old items from the cache
-func (c *Cache) Purge() PurgeStats {
-	start := time.Now()
-	var newStorage []CachedResult
-	var shortest time.Duration
-	var numPurged int
-	var timeTaken time.Duration
-
-	// Create the slice
-	newStorage = make([]CachedResult, 0)
-
-	c.storageMutex.Lock()
-
-	if len(c.Storage) > 0 {
-		// Set the shortest to a very high value
-		shortest = math.MaxInt64
+// StartPurger Starts the purge process in the background, it will be cancelled
+// when the context is cancelled. The cache will be purged initially, at which
+// point the process will sleep until the next time an item expires
+func (c *Cache) StartPurger(ctx context.Context) error {
+	c.purgeMutex.Lock()
+	if c.purgeTimer == nil {
+		c.purgeTimer = time.NewTimer(0)
+		c.purgeMutex.Unlock()
 	} else {
-		// If there is nothing in the storage then just return zero
-		shortest = 0
+		c.purgeMutex.Unlock()
+		return errors.New("purger already running")
 	}
 
-	// Loop over the storage
-	for _, ci := range c.Storage {
-		// If the item is not expired then add it to the new storage
-		if !ci.Expired() {
-			newStorage = append(newStorage, ci)
-
-			// Note the duration and store the shortest
-			if ci.CacheDuration < shortest {
-				shortest = ci.CacheDuration
-			}
-		} else {
-			// If it is expired increment the counter
-			numPurged++
-		}
-	}
-
-	// Replace the old storage with the new storage
-	c.Storage = newStorage
-	c.storageMutex.Unlock()
-
-	// Calculate the time taken
-	timeTaken = time.Since(start)
-
-	// Check of the shortest is unchanged, this can happen if all items were
-	// purged in a single purge run. In this case we need to set the shortest
-	// back to zero to avoid it being set to ~292 years
-	if shortest == math.MaxInt64 {
-		shortest = 0
-	}
-
-	return PurgeStats{
-		NumPurged:              numPurged,
-		TimeTaken:              timeTaken,
-		ShortestCacheRemaining: shortest,
-	}
-}
-
-// StartPurger Starts the automatic purging of the cache
-func (c *Cache) StartPurger() {
-	c.context, c.cancel = context.WithCancel(context.Background())
-
-	go func() {
-		var sleepTime time.Duration
-
-		// We want this process to be efficient with how often it runs. To this end
-		// we will make sure that it is able to dynamically increase and decrease
-		// its frequency depending on the duration of the caches. The logic we will
-		// use is that the purging frequency will be the shortest cache duration
-		// divided by ten. Meaning that if there is an item in the cache with a
-		// cache duration of 10 seconds, the Purger will run every second. If
-		// however this item is removed and the rest of the items are to be cached
-		// for 10 minutes, the Purger will revert to running every minute
+	go func(ctx context.Context) {
 		for {
-			timer := time.NewTimer(sleepTime)
-
 			select {
-			case <-timer.C:
-				stats := c.Purge()
+			case <-c.purgeTimer.C:
+				stats := c.Purge(time.Now())
 
-				// If the shortest cache time is zero this means that there is
-				// nothing left in the cache. In this cache we want to still sleep
-				// otherwise we would end up looping like crazy
-				if stats.ShortestCacheRemaining != 0 {
-					sleepTime = stats.ShortestCacheRemaining / 10
+				c.setNextPurgeFromStats(stats)
+			case <-ctx.Done():
+				c.purgeMutex.Lock()
+				defer c.purgeMutex.Unlock()
 
-					// Check that we aren't below the minimum wait time
-					if sleepTime < c.GetMinWaitTime() {
-						sleepTime = c.GetMinWaitTime()
-					}
-				} else {
-					sleepTime = c.GetMinWaitTime()
-				}
-
-				if stats.NumPurged > 0 {
-					log.WithFields(log.Fields{
-						"numPurged":              stats.NumPurged,
-						"timeTaken":              stats.TimeTaken,
-						"shortestCacheRemaining": stats.ShortestCacheRemaining,
-					}).Tracef("Finished %v cache purge, next purge in %v", c.Name, sleepTime)
-				}
-			case <-c.context.Done():
-				timer.Stop()
+				c.purgeTimer.Stop()
+				c.purgeTimer = nil
 				return
 			}
 		}
-	}()
+	}(ctx)
+
+	return nil
 }
 
-// StopPurger Stops the cache purging goroutine
-func (c *Cache) StopPurger() {
-	if c.cancel != nil {
-		c.cancel()
+// setNextPurgeFromStats Sets when the next purge should run based on the stats of the
+// previous purge
+func (c *Cache) setNextPurgeFromStats(stats PurgeStats) {
+	c.purgeMutex.Lock()
+	defer c.purgeMutex.Unlock()
+
+	if stats.NextExpiry == nil {
+		// If there is nothing else in the cache, wait basically
+		// forever
+		c.purgeTimer.Reset(1000 * time.Hour)
+		c.nextPurge = time.Now().Add(1000 * time.Hour)
+	} else {
+		if time.Until(*stats.NextExpiry) < c.GetMinWaitTime() {
+			c.purgeTimer.Reset(c.GetMinWaitTime())
+			c.nextPurge = time.Now().Add(c.GetMinWaitTime())
+		} else {
+			c.purgeTimer.Reset(time.Until(*stats.NextExpiry))
+			c.nextPurge = *stats.NextExpiry
+		}
+	}
+}
+
+// setNextPurgeIfNewer Sets the next time the purger will run, if the provided
+// time is sooner than the current scheduled purge time. While the purger is
+// active this will be constantly updated, however if the purger is sleeping and
+// new items are added this method ensures that the purger is woken up
+func (c *Cache) setNextPurgeIfNewer(t time.Time) {
+	c.purgeMutex.Lock()
+	defer c.purgeMutex.Unlock()
+
+	if t.Before(c.nextPurge) {
+		if c.purgeTimer == nil {
+			return
+		}
+
+		c.purgeTimer.Stop()
+		c.nextPurge = t
+		c.purgeTimer.Reset(time.Until(t))
 	}
 }
